add-all-two-localization-page

+++ src/app/pages/localization/order-status-page/order-status-page.component.ts
@@ -1,9 +1,11 @@
-import {Component, EventEmitter, OnInit, Output, ViewChild} from '@angular/core';
+import {Component, EventEmitter, Input, OnInit, Output, ViewChild} from '@angular/core';
 import {PagesService} from "../../pages.service";
 import {BasePage} from "../../@core";
 import {LocalizationServicesService} from "../services/localization-services.service";
 import {LanguageService} from "../../../modules/localization/language/language.service";
 import {IOrderStatusUp} from "../../../modules/localization/interfaces/order-status-interfaces";
+import {FormControl} from "@angular/forms";
+import {NgxUiLoaderService} from "ngx-ui-loader";
 
 @Component({
     selector: 'app-order-status-page',
@@ -12,12 +14,16 @@ import {IOrderStatusUp} from "../../../modules/localization/interfaces/order-sta
 })
 export class OrderStatusPageComponent extends BasePage implements OnInit {
     arrStatus: Array<any>
-    updateOrder: IOrderStatusUp;
     selectedOrder: any;
+    alldata:any;
+
+    public descr: FormControl = new FormControl();
+
 
     constructor(public pages: PagesService,
                 public localizationService: LocalizationServicesService,
                 public langService: LanguageService,
+                protected ngxService: NgxUiLoaderService,
     ) {
         super(pages);
 
@@ -28,8 +34,6 @@ export class OrderStatusPageComponent extends BasePage implements OnInit {
         super.initPanelButton();
         this.getStatus();
         this.getLangList();
-        console.log(this.selectedOrder)
-        // this.getList();
     }
 
 
@@ -45,41 +49,54 @@ export class OrderStatusPageComponent extends BasePage implements OnInit {
     };
 
     getStatus(): void {
-        this.localizationService.getOrderStatus().subscribe(data => {
+        this.localizationService.getOrderStatus().subscribe(
+            data => {
             this.arrStatus = data.data;
-            console.log(this.arrStatus)
+            this.alldata=data;
+            this.localizationService.data=data;
+            console.log( this.alldata)
         })
+
     }
 
     deleteStatus(order): void {
-        this.localizationService.deleteOrderStatus(order.id).subscribe(data => {
-            this.arrStatus = data.data;
-        })
-        this.getStatus()
+        if(order.name!=='New') {
+            this.localizationService.deleteOrderStatus(order.id).subscribe(data => {
+                this.getStatus()
+            })
+        }
+        else {
+            console.log('you cant delete this status ')
+        }
+
     }
 
-    edit(item) {
-        this.selectedOrder = item;
+    edit(i) {
+        this.selectedOrder = i;
         this.openForm();
+
     }
 
     save = () => {
         const updateOrder = {
             name: this.selectedOrder.descriptions[0].dectiption,
-            description: this.selectedOrder.descriptions
+            description: this.selectedOrder.descriptions,
         }
+        console.log(updateOrder)
         if (this.selectedOrder.id !== undefined) {
             this.localizationService.editOrderStaus(this.selectedOrder.id, updateOrder).subscribe(data => {
             })
+            this.getStatus()
         }
         else {
             this.localizationService.addNewOrderStatus(this.selectedOrder.id, updateOrder).subscribe(data => {
+                this.getStatus()
+
             })
 
         }
-        console.log(this.selectedOrder)
-        this.getStatus()
         this.closeForm();
+
 
     }
     plus = () => {
@@ -87,6 +104,37 @@ export class OrderStatusPageComponent extends BasePage implements OnInit {
         this.selectedOrder = this.localizationService.selectedOrder;
         this.openForm();
     };
+    postHandler = (data) => {
+        // this.ngxService.stopAll();
+        this.localizationService.data.data.push(data.data);
+        this.localizationService.data.count++;
+        this.closeForm();
+        // this.toastr.success("option ADDED");
+    };
+    getListHandler = (data) => {
+        this.ngxService.stopAll();
+        this.localizationService.data = data;
+    };
+
+    //#region pagination
+
+    pageToHandler(page: number): void {
+        this.localizationService.page = page;
+    }
+    pagePrevHandler(): void {
+        this.localizationService.page--;
+    }
+    pageNextHandler(): void {
+        this.localizationService.page++;
+    }
+    pageChangedHandler(): void {
+       this.getStatus();;
+        window.scrollTo(0, 0);
+    }
+    Math = Math;
+
+//#endregion
+
 
 
 }
